Rework geometry handling, modify scoring volume and information
The detector geometry is modifed to use G4SubtractionSolid for the
teflon box and the copper piece. A solid+logical volume of air is
added to represent the PMT. This logical volume is used as the
scoring volume.

The G4Accumulables and methods of the TestSimRunActions and
TestSimEventActions have been renamed and modified to reflect the
counting of PMT fires.