Tie photon source position to copper piece
The positioning of the particle gun is now tied to wherever the copper
piece is positioned - it will always be placed immediately in front of
the copper, facing into the box. This paves the way for i.e. moving the
PMT-copper assembly forward and backward, as well as rotating it about
the z-axis, without any inconsistency in the placement of the photon
source.