Model LED spectral energy spread
The particle gun now samples a Gaussian with FWHM corresponding to 12nm
about a central wavelength of 290nm. This currently renders the
RunAction status line regarding the energy of the particles fire
inaccurate, as it will simply report whichever energy the last
opticalphoton was given. Modifying the RunAction so it prints the
average particle energy may be preferable.