Add EventAction volume members, track additional physics info, modify det. const.
Several new fields are added to the EventAction class in order to access the
detector's world and photon catcher volumes, which are necessary for extracting
physics information in SteppingAction. SteppingAction tests to see if an incident
photon is above 300 nm (via its energy equivalent) before it fires PMT. It also
tests if the photon is in the world volume, which indicates a scattering in air.
The behavior is unexpected, so this assumption may be invalid.

DetectorConstruction is modified to include a box containing the entire assembly
which will absorb all photons that escape the box past the PMT. The world volume
and assembly (PMT+copper+box) position is modified so that the PMT face is no
longer located at Z=0.