Correct modeling of LED spectral distribution
The previous model of the LED spectral energy distribution erroneously
assumed that the distribution of the energy would be, like the spread in
wavelength, Gaussian. However, as the energy is a function of the
reciprocal of the wavelength, this is not the case. The energy is now
calculated using wavelength values sampled from a Gaussian distribution.