Rework analysis data handling, save primary kinematics, add branches
The branch handling has been moved from EventAction
to AnalysisManager, which allows the AnalysisManager to be accessed
from the PrimaryGeneratorAction class. This allows the primary kinematics
to be saved to the tree.

Other branches have been added to the tree, namely the number of reflections
and absorptions of the copper piece. Also added are hyperparameters
PMTinsertion and CopperRotation, which will require further reworking of
TestSim.cc to implement. As is, it is not possible for the DetectorConstruction
to access the AnalysisManager, which would be required in order for these
branches to be filled. But this may not be the desired route and these
branches may ultimately removed.

A countably infinite number of minor changes have also been made.